pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kajovospend"
version = "0.1.0"
description = "KájovoSpend desktop application"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "OpenAI"}]
dependencies = [
  "PySide6>=6.8,<6.9",
  "SQLAlchemy>=2.0,<2.1",
  "pydantic>=2.8,<3",
  "platformdirs>=4.2,<5",
  "requests>=2.32,<3",
  "keyring>=25.6,<26",
  "openpyxl>=3.1,<4",
  "Pillow>=11,<12",
  "pypdf>=5,<6",
  "pypdfium2>=4.30,<6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-qt>=4.4,<5",
  "ruff>=0.6,<1",
  "mypy>=1.11,<2",
  "pyinstaller>=6.10,<7",
]

[project.scripts]
kajovospend = "kajovospend.app.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["kajovospend*"]
exclude = ["tests*"]
namespaces = false

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
